+++ R/tools.R
@@ -547,6 +547,82 @@ test_e2e <- function(interactive = FALSE) {
   }
 }
 
+#' Run a unit test coverage check
+#'
+#' Uses the `{covr}` package to produce unit test coverage reports.
+#' Uses the `{testhat}` package to run all unit tests in `tests/testthat` directory.
+#'
+#' @param source_files Character vector of source files with function definitions to measure
+#' coverage. Defaults to all `.R` files in the `app` tree.
+#' @param test_files Character vector of test files with code to test the functions. Defaults to
+#'   all test files in `tests/testthat` with the `test-<name>.R` filename pattern.
+#' @param line_exclusions passed to `covr::file_coverage`
+#' @param function_exclusions passed to `covr::file_coverage`
+#' @return A `covr` coverage dataset.
+#'
+#' @examples
+#' if (interactive()) {
+#'   # Run a test coverage check for the entire rhino app
+#'   # using all tests in the `tests/testthat` directory.
+#'   covr_r()
+#' }
+#'
+#' @export
+covr_r <- function(
+    source_files = list.files("app",
+                              pattern = "\\.[rR]$",
+                              full.names = TRUE,
+                              recursive = TRUE),
+    test_files = list.files("tests/testthat",
+                            pattern = "^test-.*\\.R",
+                            full.names = TRUE,
+                            recursive = TRUE),
+    line_exclusions = NULL,
+    function_exclusions = NULL) {
+
+  withr::with_file("box_loader.R", {
+    module_list <- sub(
+      "__init__",
+      "`__init__`",
+      paste0(tools::file_path_sans_ext(source_files), ",")
+    )
+
+    loader_lines <- c("box::use(", module_list, ")")
+
+    writeLines(loader_lines, "box_loader.R")
+
+    coverage <- covr::file_coverage(
+      source_files = "box_loader.R",
+      test_files = test_files,
+      line_exclusions = line_exclusions,
+      function_exclusions = function_exclusions
+    )
+  })
+
+  coverage
+}
+
+#' Display rhino test coverage results using a standalone report
+#'
+#' Uses the `{covr}` package to produce unit test coverage reports.
+#' Uses the `{testhat}` package to run all unit tests in `tests/testthat` directory.
+#'
+#' @param rhino_coverage a rhino coverage dataset, defaults to `covr_r()`.
+#' @param ... additional arguments to pass to
+#'        [`covr::report()`](https://covr.r-lib.org/reference/report.html)
+#' @return None. This function is called for side effects.
+#'
+#' @examples
+#' if (interactive()) {
+#'   # Run a test coverage report on a rhino app
+#'   covr_report()
+#' }
+#'
+#' @export
+covr_report <- function(rhino_coverage = covr_r(), ...) {
+  covr::report(x = rhino_coverage, ...)
+}
+
 #' Development mode
 #'
 #' Run application in development mode with automatic rebuilding and reloading.